// src/api/openapi/models/AppBusinessConfigResponseDto.ts
export interface AppBusinessConfigResponseDto {
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    minEndOrderFromNowInHours: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    minDateToAcceptOfferInHours: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    minDiffBetweenStartTripAndEndOrderForAllPaymentsInHours: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    minDiffBetweenStartTripAndEndOrderInHours: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    payInvoiceWarningAfterAcceptOfferInHours: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    payRestOfPriceWarningBeforeStartTripInHours: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripDepositInPercentage: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelFeePercentageForDemander: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelFeeAfterLimitPercentageForDemander: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripOfferCommissionPercentage: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelPenaltyLimitInDays: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelPenaltyPercentageForTransporterFromCompany: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelPenaltyMinAmountInCzkForTransporterFromCompany: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelPenaltyPercentageForTransporterFromDemander: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelPenaltyMinAmountInCzkForTransporterFromDemander: number;
    /**
     * 
     * @type {number}
     * @memberof AppBusinessConfigResponseDto
     */
    tripCancelPenaltyAfterLimitPercentageForTransporter: number;
}